Move var methods to J8Spec facade
