+++ src/main/java/j8spec/J8Spec.java
@@ -353,6 +353,43 @@ public final class J8Spec {
     }
 
     /**
+     * Creates a wrapper object to allow "final" variables to have their value modified. The initial
+     * value is <code>null</code>.
+     *
+     * @param <T> type of value the variable object can store
+     * @return new variable object
+     * @since 3.1.0
+     */
+    public static <T> Var<T> var() {
+        return new Var<>();
+    }
+
+    /**
+     * Access the value stored in the given variable object.
+     *
+     * @param var variable object
+     * @param <T> type of value the variable object can store
+     * @return value stored in the variable object
+     * @since 3.1.0
+     */
+    public static <T> T var(Var<T> var) {
+        return var.value;
+    }
+
+    /**
+     * Stores the given value in the provided variable object.
+     *
+     * @param var variable object
+     * @param value value to be stored
+     * @param <T> type of value the variable object can store
+     * @return value stored in the variable object
+     * @since 3.1.0
+     */
+    public static <T> T var(Var<T> var, T value) {
+        return var.value = value;
+    }
+
+    /**
      * Uses the given spec class to build and populate a list of {@link Example} objects ready to be executed.
      *
      * @param specClass class with a public default constructor that contains the spec definition
