+++ src/test/java/j8spec/VarTest.java
@@ -2,7 +2,7 @@ package j8spec;
 
 import org.junit.Test;
 
-import static j8spec.Var.var;
+import static j8spec.J8Spec.var;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
